BottomNavigationbar move MainActivity -> BoxstoreMenuActivity
